Fix FontImage jsdocs
I noticed that the type for the param `image` was of type `new (width?: number, height?: number) => HTMLImageElement`. This doesn't make sense. I then realized this was because in the jsdocs for this class, the type was set as `Image`. In reality, the line `new Image(...)` returns an instance of `HTMLImageElement` so that's what the jsdocs should say.

Sorry, this is my first contribution to this project, and there's no contribution guidelines. Hope this PR makes sense.