Keyboard input uses strings
Input has been switched over to use the code string. The old way of using the keyCode has been deprecated, and was not a great way to do it. This is the way that is recommended for games.

For anyone using keyboard input, you will need to replace calls to isKeyDown and others with the keyboard key like "KeyW" or whatever it is. 🙏