make drawText draw to mainCanvas by default
it is a little confusing that by default the drawText draws to overlay while nothing else does. that is because it is an ugly function call to change the target canvas.

LittleJS.drawText(this.text, this.pos, 1, hsl(0, 0, 1), undefined, undefined, undefined, undefined, LittleJS.mainContext);

So instead we will make drawText just draw to the main context by default and add a new helper function drawOverlay that wraps up that call to draw it to the overlay.

The context param has also beem moved to be after the max length since it is no longer as useful.